django/requirements.txt
asgiref==3.3.4
coverage==5.5
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
numpy==1.21.4
pandas==1.3.4
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
